fix: bug with modals ( with react-onclickoutside)
