+++ package.json
@@ -16,6 +16,7 @@
     "react-helmet": "^5.2.0",
     "react-intl": "^2.7.2",
     "react-intl-formatted-duration": "^3.0.0",
+    "react-onclickoutside": "^6.7.1",
     "react-redux": "^6.0.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.1",
