+++ src/ui/SearchFilters/PassengersPicker.js
@@ -1,7 +1,7 @@
 import React, { Component, Fragment } from "react";
 import "./PassengersPicker.sass";
-import Checkbox from "../../Search/Filters/Checkbox";
 import pluralize from "pluralize-ru";
+import Content from "./PassengersChoose";
 class PassengersPicker extends Component {
   state = {
     isOpened: false,
@@ -21,6 +21,9 @@ class PassengersPicker extends Component {
     this.setState(prevState => {
       return { isOpened: !prevState.isOpened };
     });
+  };
+  disableContent = () => {
+    this.setState({ isOpened: false });
   };
   decrementAdults = () => {
     this.state.adults > 1 &&
@@ -111,6 +114,7 @@ class PassengersPicker extends Component {
                 incrementKids={this.incrementKids}
                 incrementToddlers={this.incrementToddlers}
                 setBuisness={this.setBuisness}
+                disableContent={this.disableContent}
               />
             </div>
           )}
@@ -121,88 +125,3 @@ class PassengersPicker extends Component {
 }
 
 export default PassengersPicker;
-
-class Content extends Component {
-  state = {};
-
-  render() {
-    return (
-      <Fragment>
-        <div className="content__wrapper">
-          <div className="content__item">
-            <div className="content__item__text">Взрослые</div>
-            <div className="content__item__button">
-              <button
-                onClick={this.props.decrementAdults}
-                className="content__item__button__decrement"
-              >
-                -
-              </button>
-              <div className="content__item__button__count">
-                {this.props.adults}
-              </div>
-              <button
-                onClick={this.props.incrementAdults}
-                className="content__item__button__increment"
-              >
-                +
-              </button>
-            </div>
-          </div>
-          <div className="content__item">
-            <div className="content__item__text">Дети до 12 лет</div>
-            <div className="content__item__button">
-              {" "}
-              <button
-                onClick={this.props.decrementKids}
-                className="content__item__button__decrement"
-              >
-                -
-              </button>
-              <div className="content__item__button__count">
-                {this.props.kids}
-              </div>
-              <button
-                onClick={this.props.incrementKids}
-                className="content__item__button__increment"
-              >
-                +
-              </button>
-            </div>
-          </div>
-          <div className="content__item">
-            <div className="content__item__text">
-              Дети до 2 лет
-              <p className="content__item__subtext">Без места</p>
-            </div>
-            <div className="content__item__button">
-              {" "}
-              <button
-                onClick={this.props.decrementToddlers}
-                className="content__item__button__decrement"
-              >
-                -
-              </button>
-              <div className="content__item__button__count">
-                {this.props.toddlers}
-              </div>
-              <button
-                onClick={this.props.incrementToddlers}
-                className="content__item__button__increment"
-              >
-                +
-              </button>
-            </div>
-          </div>
-        </div>
-        <div>
-          <Checkbox>
-            <div onClick={this.props.setBuisness} style={{ marginTop: 2 }}>
-              Бизнес-класс
-            </div>
-          </Checkbox>
-        </div>
-      </Fragment>
-    );
-  }
-}
